Move helper Pathname methods into Refinement
https://docs.ruby-lang.org/en/2.4.0/syntax/refinements_rdoc.html
